requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
SQLAlchemy
psycopg2-binary